to long

+++ src/com/ipca/GregoryLeibniz.java
@@ -67,7 +67,7 @@ public class GregoryLeibniz {
 		double factor = 1.0;
 		double sum = 0.0;
 
-		for (int k = 0; k < numberTotalGenerated; k++) {
+		for (long k = 0; k < numberTotalGenerated; k++) {
 			sum += factor / (2 * k + 1);
 			factor = -factor;
 		}
